backend/main/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
langgraph>=0.3.0
# langgraph-checkpoint-postgres — uncomment for production persistence:
# pip install langgraph-checkpoint-postgres
# Then swap InMemorySaver in graph/pipeline.py for AsyncPostgresSaver.
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
google-genai>=1.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
boto3>=1.34.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
python-json-logger>=2.0.7

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
httpx>=0.27.0
